More common functions CSprite2d for Vice City

+++ plugin_sa/game_sa/CBike.h
@@ -43,7 +43,7 @@ public:
     CColPoint      m_anWheelColPoint[4];
     float field_710[4];
     float field_720[4];
-    int field_730[4]; // possibly wheel rotation speed?
+    float field_730[4]; // possibly wheel rotation speed?
     float field_740;
     int            m_anWheelSurfaceType[2];
     char field_74C[2];
